chore(net,connmanager): move to internal
