+++ internal/api/api.go
@@ -14,7 +14,7 @@ import (
 	"nimona.io/pkg/http/router"
 	"nimona.io/pkg/keychain"
 	"nimona.io/pkg/log"
-	"nimona.io/pkg/net"
+	"nimona.io/internal/net"
 	"nimona.io/pkg/object"
 	"nimona.io/pkg/objectmanager"
 	"nimona.io/pkg/resolver"
