+++ internal/connmanager/connmanager_test.go
@@ -6,11 +6,10 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
+	"nimona.io/internal/net"
 	"nimona.io/pkg/context"
 	"nimona.io/pkg/crypto"
-	"nimona.io/pkg/eventbus"
 	"nimona.io/pkg/keychain"
-	"nimona.io/pkg/net"
 	"nimona.io/pkg/object"
 	"nimona.io/pkg/peer"
 )
@@ -22,16 +21,16 @@ func TestGetConnection(t *testing.T) {
 		return nil
 	}
 
-	eb1, _, n1 := newPeer(t)
-	eb2, kc2, n2 := newPeer(t)
+	_, n1 := newPeer(t)
+	kc2, n2 := newPeer(t)
 
-	mgr := New(ctx, eb1, n1, handler)
+	mgr := New(ctx, n1, handler)
 
 	lst1, err := n1.Listen(ctx, "0.0.0.0:0")
 	assert.NoError(t, err)
 	defer lst1.Close()
 
-	mgr2 := New(ctx, eb2, n2, handler)
+	mgr2 := New(ctx, n2, handler)
 	assert.NotNil(t, mgr2)
 
 	lst2, err := n2.Listen(ctx, "0.0.0.0:0")
@@ -61,7 +60,6 @@ func TestGetConnection(t *testing.T) {
 }
 
 func newPeer(t *testing.T) (
-	eventbus.Eventbus,
 	keychain.Keychain,
 	net.Network,
 ) {
@@ -70,13 +68,10 @@ func newPeer(t *testing.T) (
 	pk, err := crypto.GenerateEd25519PrivateKey()
 	assert.NoError(t, err)
 
-	eb := eventbus.New()
-
 	kc := keychain.New()
-	kc.Put(keychain.PrimaryPeerKey, pk)
+	kc.PutPrimaryPeerKey(pk)
 
-	return eb, kc, net.New(
-		net.WithKeychain(kc),
-		net.WithEventBus(eb),
+	return kc, net.New(
+		kc,
 	)
 }
