+++ internal/daemon/daemon.go
@@ -16,7 +16,7 @@ import (
 	"nimona.io/pkg/exchange"
 	"nimona.io/pkg/keychain"
 	"nimona.io/pkg/nat"
-	"nimona.io/pkg/net"
+	"nimona.io/internal/net"
 	"nimona.io/pkg/object"
 	"nimona.io/pkg/objectmanager"
 	"nimona.io/pkg/peer"
