+++ internal/net/connection.go
