+++ pkg/exchange/exchange_test.go
@@ -13,7 +13,7 @@ import (
 	"nimona.io/pkg/crypto"
 	"nimona.io/pkg/eventbus"
 	"nimona.io/pkg/keychain"
-	"nimona.io/pkg/net"
+	"nimona.io/internal/net"
 	"nimona.io/pkg/object"
 	"nimona.io/pkg/peer"
 )
