+++ playground/certificate-request/serve.go
@@ -12,7 +12,7 @@ import (
 	"nimona.io/pkg/crypto"
 	"nimona.io/pkg/eventbus"
 	"nimona.io/pkg/keychain"
-	"nimona.io/pkg/net"
+	"nimona.io/internal/net"
 	"nimona.io/pkg/object"
 	"nimona.io/pkg/peer"
 	"nimona.io/pkg/resolver"
